event_processing/requirements.txt
kafka-python
Flask
mysql-connector-python
